+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Description list component (#39)
 - Drawer component (#42)
 - Empty state component (#44)
+- Hint component (#56)
 - Panel component (#77) - thanks @halkosajtarevic
 - Tabs components (#91)